+++ services-custom/dynamodb-enhanced/src/main/java/software/amazon/awssdk/enhanced/dynamodb/mapper/StaticTableSchema.java
@@ -31,6 +31,7 @@ import java.util.function.Function;
 import java.util.function.Supplier;
 import java.util.stream.Stream;
 import software.amazon.awssdk.annotations.SdkPublicApi;
+import software.amazon.awssdk.enhanced.dynamodb.Attribute;
 import software.amazon.awssdk.enhanced.dynamodb.AttributeConverter;
 import software.amazon.awssdk.enhanced.dynamodb.AttributeConverterProvider;
 import software.amazon.awssdk.enhanced.dynamodb.DefaultAttributeConverterProvider;
@@ -71,9 +72,10 @@ import software.amazon.awssdk.services.dynamodb.model.AttributeValue;
  */
 @SdkPublicApi
 public final class StaticTableSchema<T> implements TableSchema<T> {
-    private final List<ResolvedStaticAttribute<T>> attributeMappers;
+
+    private final List<ResolvedStaticAttribute<T, ?>> attributeMappers;
     private final Supplier<T> newItemSupplier;
-    private final Map<String, ResolvedStaticAttribute<T>> indexedMappers;
+    private final Map<String, ResolvedStaticAttribute<T, ?>> indexedMappers;
     private final StaticTableMetadata tableMetadata;
     private final EnhancedType<T> itemType;
     private final AttributeConverterProvider attributeConverterProvider;
@@ -85,12 +87,12 @@ public final class StaticTableSchema<T> implements TableSchema<T> {
                 ConverterProviderResolver.resolveProviders(builder.attributeConverterProviders);
 
         // Resolve declared attributes and find converters for them
-        Stream<ResolvedStaticAttribute<T>> attributesStream = builder.attributes == null ?
+        Stream<ResolvedStaticAttribute<T, ?>> attributesStream = builder.attributes == null ?
             Stream.empty() : builder.attributes.stream().map(a -> a.resolve(this.attributeConverterProvider));
 
         // Merge resolved declared attributes and additional attributes that were added by extend or flatten
-        List<ResolvedStaticAttribute<T>> mutableAttributeMappers = new ArrayList<>();
-        Map<String, ResolvedStaticAttribute<T>>  mutableIndexedMappers = new HashMap<>();
+        List<ResolvedStaticAttribute<T, ?>> mutableAttributeMappers = new ArrayList<>();
+        Map<String, ResolvedStaticAttribute<T, ?>>  mutableIndexedMappers = new HashMap<>();
         Stream.concat(attributesStream, builder.additionalAttributes.stream()).forEach(
             resolvedAttribute -> {
                 String attributeName = resolvedAttribute.attributeName();
@@ -136,7 +138,7 @@ public final class StaticTableSchema<T> implements TableSchema<T> {
      */
     public static final class Builder<T> {
         private final Class<T> itemClass;
-        private final List<ResolvedStaticAttribute<T>> additionalAttributes = new ArrayList<>();
+        private final List<ResolvedStaticAttribute<T, ?>> additionalAttributes = new ArrayList<>();
 
         private List<StaticAttribute<T, ?>> attributes;
         private Supplier<T> newItemSupplier;
@@ -243,7 +245,7 @@ public final class StaticTableSchema<T> implements TableSchema<T> {
          * the super-class into the {@link StaticTableSchema} of the sub-class.
          */
         public Builder<T> extend(StaticTableSchema<? super T> superTableSchema) {
-            Stream<ResolvedStaticAttribute<T>> attributeStream =
+            Stream<ResolvedStaticAttribute<T, ?>> attributeStream =
                 upcastingTransformForAttributes(superTableSchema.attributeMappers);
             attributeStream.forEach(this.additionalAttributes::add);
             return this;
@@ -335,8 +337,8 @@ public final class StaticTableSchema<T> implements TableSchema<T> {
             return new StaticTableSchema<>(this);
         }
 
-        private static <T extends R, R> Stream<ResolvedStaticAttribute<T>> upcastingTransformForAttributes(
-            Collection<ResolvedStaticAttribute<R>> superAttributes) {
+        private static <T extends R, R> Stream<ResolvedStaticAttribute<T, ?>> upcastingTransformForAttributes(
+            Collection<ResolvedStaticAttribute<R, ?>> superAttributes) {
             return superAttributes.stream().map(attribute -> attribute.transform(x -> x, null));
         }
     }
@@ -355,7 +357,7 @@ public final class StaticTableSchema<T> implements TableSchema<T> {
             String key = entry.getKey();
             AttributeValue value = entry.getValue();
             if (!isNullAttributeValue(value)) {
-                ResolvedStaticAttribute<T> attributeMapper = indexedMappers.get(key);
+                Attribute<T, ?> attributeMapper = indexedMappers.get(key);
 
                 if (attributeMapper != null) {
                     if (item == null) {
@@ -403,12 +405,7 @@ public final class StaticTableSchema<T> implements TableSchema<T> {
 
     @Override
     public AttributeValue attributeValue(T item, String key) {
-        ResolvedStaticAttribute<T> attributeMapper = indexedMappers.get(key);
-
-        if (attributeMapper == null) {
-            throw new IllegalArgumentException(String.format("TableSchema does not know how to retrieve requested "
-                                                             + "attribute '%s' from mapped object.", key));
-        }
+        Attribute<T, ?> attributeMapper = attribute(key);
 
         AttributeValue attributeValue = attributeMapper.attributeGetterMethod().apply(item);
 
@@ -418,6 +415,17 @@ public final class StaticTableSchema<T> implements TableSchema<T> {
     @Override
     public EnhancedType<T> itemType() {
         return this.itemType;
+    }
+
+    @Override
+    public Attribute<T, ?> attribute(String key) {
+        ResolvedStaticAttribute<T, ?> attributeMapper = indexedMappers.get(key);
+
+        if (attributeMapper == null) {
+            throw new IllegalArgumentException(String.format("TableSchema does not know how to retrieve requested "
+                    + "attribute '%s' from mapped object.", key));
+        }
+        return attributeMapper;
     }
 
     /**