+++ services-custom/dynamodb-enhanced/src/test/java/software/amazon/awssdk/enhanced/dynamodb/mapper/StaticAttributeTest.java
@@ -215,7 +215,7 @@ public class StaticAttributeTest {
                                                                          .attributeConverter(attributeConverter)
                                                                          .build();
 
-        ResolvedStaticAttribute<SimpleItem> resolvedAttribute =
+        ResolvedStaticAttribute<SimpleItem, ?> resolvedAttribute =
             staticAttribute.resolve(AttributeConverterProvider.defaultProvider());
 
         Function<SimpleItem, AttributeValue> attributeValueFunction = resolvedAttribute.attributeGetterMethod();